requirements.txt
Django~=3.0
celery>=4.4.2
requests>=2.23.0
pytesseract>=0.3.3
Pillow>=7.0.0
redis
gunicorn
whitenoise
future
boto3
django-storages
dj-database-url
psycopg2
python-dotenv